adds callback support.
I also added a script to run LVGL inside of wxPython as well. I have not tested this 100% yet.

It now compiles and runs on Windows as well.

I had to heavily modify LVGL in order to get the callback functions to work properly. I am pretty sure there is a better way to go about doing what I did.