removes dynamic creation of python objects
There is no more stub file, the build script generates a python source file that creates objects properly. I also figured out the arrays and being able to pass functions directly as callbacks without having to wrap them in a python object first. I think I can improve the way the arrays are handled and I am going to mess about with that some.

The LVGL that is included has the performance monitor fixed so it reports a correct value and I also added a callback for incrementing the tick. There is an example of the tick callback in the examples folder. Having that callback eliminates the need to use an ISR all together allowing a user to have the ability to call the task handler whenever they want in their code because the only time incrementing the tick really matters is when the task handler has been called. This allows the task handler to get the most accurate time while it is running without an ISR having to stop the task handler in order to increment the tick.